Revert some changes from previous 2 commits
Affine example's "bold upright" line renders as 2 lines regardless.
Showing right-hand side (RHS) of assignment on the next line is a
syntactic error. By adding spacing, part of RHS stays on the same line,
therefore showing non-error code.

"In parallel" is not used as a generic introductory phrase.

"As is" is written with a space, even in FOSS licenses.

The website is "typst.app", not "Typst.app", and isn't advertised as
such.

It's called commutative "diagram", not "layout".

The period in caption is automatic.

The @fig:tree reference is out of place. It will be included later.

It's "recursive decent parser".

"Docs" is not a formal abbreviation, I assume.